LoadingMixin plus multiple small UI bugfixes
- Implementation of LoadingMixin
- Update of all (?) components that load data to using the LoadingMixin

The LoadingMixin handles loading state, i.e. when a component triggers an asynchronous reload of data it can call this.loading() to set the components state to loading. Then, in the render function calling this.isLoading() it can check if the state is currently loading, allowing elements to e.g., be disabled during load (like buttons). Once the AJAX call returns the component using the mixin should call this.loaded({ data : data }) where 'data' could be any state field the component uses for whatever it is displaying. The call will then update the state of the component to non-loading and merge in the data passed to the components state.